refactor pytests to follow structure
