+++ tests/apps/tower/test_tower_api.py
@@ -4,10 +4,10 @@ from typing import List
 
 import pytest
 
+from tests.apps.tower.conftest import TowerResponseFile
 from tests.mocks.tower_mock import MockTowerAPI
-from tests.store.utils.conftest import TowerResponseFile, TowerTaskResponseFile
+from trailblazer.apps.tower.models import TowerTask
 from trailblazer.constants import TrailblazerStatus
-from trailblazer.store.models import Job
 
 
 @pytest.mark.parametrize(
@@ -145,3 +145,27 @@ def test_tower_api_tasks_empty(
 
     # THEN an empty list should be returned
     assert jobs == []
+
+
+def test_tower_task_properties(
+    tower_task: TowerTask,
+    created_at,
+    started_at,
+    last_updated,
+    tower_task_duration,
+    slurm_id,
+    tower_task_name,
+) -> None:
+    """Assess that TowerTask returns the right properties."""
+
+    # GIVEN a tower task
+
+    # THEN properties should be returned
+    assert tower_task.process == tower_task_name
+    assert tower_task.nativeId == slurm_id
+    assert tower_task.status == TrailblazerStatus.COMPLETED.value
+    assert tower_task.dateCreated == created_at
+    assert tower_task.lastUpdated == last_updated
+    assert tower_task.start == started_at
+    assert tower_task.duration == tower_task_duration
+    assert tower_task.is_complete is True
